+++ src/pages/adminPages/AdminTestingFeedback.tsx
@@ -3,6 +3,11 @@ import { FaBug, FaPlus, FaSearch, FaSort, FaEye, FaEdit, FaTrash, FaThumbsUp, Fa
 import { FeedbackService } from '../../services/feedbackService';
 import { FeedbackTicket, FeedbackStats, FeedbackFilters, FeedbackStatus, FeedbackPriority, FeedbackCategory } from '../../types/feedback';
 import { useAuth } from '../../hooks/useAuth';
+import CreateTicketModal from '../../components/modals/CreateTicketModal';
+import EditTicketModal from '../../components/modals/EditTicketModal';
+import ViewTicketModal from '../../components/modals/ViewTicketModal';
+import CommentsSidebar from '../../components/modals/CommentsSidebar';
+import DeleteTicketModal from '../../components/modals/DeleteTicketModal';
 import '../../styles/adminStyles/AdminTestingFeedback.css';
 
 const STATUS_OPTIONS: FeedbackStatus[] = ['open', 'in_progress', 'resolved', 'closed', 'duplicate'];
@@ -83,48 +88,20 @@ export default function AdminTestingFeedback() {
   const [stats, setStats] = useState<FeedbackStats | null>(null);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
-  const [selectedTicket, setSelectedTicket] = useState<FeedbackTicket | null>(null);
-  const [showTicketModal, setShowTicketModal] = useState(false);
+  // Modal states
+  const [showCreateModal, setShowCreateModal] = useState(false);
   const [showEditModal, setShowEditModal] = useState(false);
   const [editingTicket, setEditingTicket] = useState<FeedbackTicket | null>(null);
+  const [showViewModal, setShowViewModal] = useState(false);
+  const [viewingTicket, setViewingTicket] = useState<FeedbackTicket | null>(null);
   const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [deletingTicket, setDeletingTicket] = useState<FeedbackTicket | null>(null);
-  const [showCreateModal, setShowCreateModal] = useState(false);
-  const [newTicket, setNewTicket] = useState({
-    title: '',
-    description: '',
-    category: 'bug' as FeedbackCategory,
-    priority: 'medium' as FeedbackPriority,
-    tags: [] as string[],
-    tagInput: '',
-    attachments: [] as File[],
-    // Testing-specific fields
-    urlToPage: '',
-    browser: '',
-    browserVersion: '',
-    operatingSystem: '',
-    deviceType: 'desktop' as 'desktop' | 'mobile' | 'tablet',
-    screenResolution: '',
-    stepsToReproduce: '',
-    expectedBehavior: '',
-    actualBehavior: '',
-    severity: 'minor' as 'cosmetic' | 'minor' | 'major' | 'critical' | 'blocker',
-    environment: 'production' as 'development' | 'staging' | 'production',
-    testCaseId: '',
-    regression: false,
-    workaround: ''
-  });
+  const [showCommentsSidebar, setShowCommentsSidebar] = useState(false);
+  const [selectedTicketForComments, setSelectedTicketForComments] = useState<FeedbackTicket | null>(null);
   const [filters, setFilters] = useState<FeedbackFilters>({});
   const [searchTerm, setSearchTerm] = useState('');
   const [sortBy, setSortBy] = useState<'createdAt' | 'updatedAt' | 'priority' | 'votes'>('createdAt');
   const [sortOrder, setSortOrder] = useState<'asc' | 'desc'>('desc');
-  const [showCommentsSidebar, setShowCommentsSidebar] = useState(false);
-  const [selectedTicketForComments, setSelectedTicketForComments] = useState<FeedbackTicket | null>(null);
-  const [newComment, setNewComment] = useState({
-    content: '',
-    isInternal: false,
-    attachments: [] as File[]
-  });
   const [actionLoading, setActionLoading] = useState<Record<string, boolean>>({});
   const [hoveredTicket, setHoveredTicket] = useState<string | null>(null);
   const [selectedCardRef, setSelectedCardRef] = useState<HTMLDivElement | null>(null);
@@ -208,26 +185,6 @@ export default function AdminTestingFeedback() {
     loadTickets();
   }, [loadTickets]);
 
-  // const handleStatusChange = async (ticketId: string, newStatus: FeedbackStatus) => {
-  //   try {
-  //     await FeedbackService.updateTicket(ticketId, { status: newStatus });
-  //     await loadTickets();
-  //   } catch (err) {
-  //     console.error('Error updating ticket status:', err);
-  //     setError('Failed to update ticket status');
-  //   }
-  // };
-
-  // const handlePriorityChange = async (ticketId: string, newPriority: FeedbackPriority) => {
-  //   try {
-  //     await FeedbackService.updateTicket(ticketId, { priority: newPriority });
-  //     await loadTickets();
-  //   } catch (err) {
-  //     console.error('Error updating ticket priority:', err);
-  //     setError('Failed to update ticket priority');
-  //   }
-  // };
-
   const handleVote = async (ticketId: string, voteType: 'up' | 'down') => {
     const userId = userProfile?.uid || '';
     const actionKey = `vote-${ticketId}-${voteType}`;
@@ -303,105 +260,53 @@ export default function AdminTestingFeedback() {
     }
   };
 
-
-  const handleCreateTicket = async () => {
-    if (!newTicket.title.trim() || !newTicket.description.trim()) {
-      setError('Please fill in all required fields');
-      return;
-    }
-
+  const handleCreateTicket = async (ticketData: any) => {
     // Create optimistic ticket
     const optimisticTicket: FeedbackTicket = {
       id: `temp-${Date.now()}`,
       sequentialId: 0, // Will be updated when real ticket is created
-      title: newTicket.title.trim(),
-      description: newTicket.description.trim(),
-      category: newTicket.category,
-      priority: newTicket.priority,
+      title: ticketData.title.trim(),
+      description: ticketData.description.trim(),
+      category: ticketData.category,
+      priority: ticketData.priority,
       status: 'open',
       reporterId: userProfile?.uid || '',
       reporterName: userProfile?.displayName || 'Unknown User',
       reporterEmail: userProfile?.email || '',
       createdAt: new Date(),
       updatedAt: new Date(),
-      tags: newTicket.tags,
+      tags: ticketData.tags,
       attachments: [],
       comments: [],
       votes: 0,
       upvoters: [],
       downvoters: [],
       // Testing-specific fields
-      urlToPage: newTicket.urlToPage.trim() || undefined,
-      browser: newTicket.browser.trim() || undefined,
-      browserVersion: newTicket.browserVersion.trim() || undefined,
-      operatingSystem: newTicket.operatingSystem.trim() || undefined,
-      deviceType: newTicket.deviceType,
-      screenResolution: newTicket.screenResolution.trim() || undefined,
-      stepsToReproduce: newTicket.stepsToReproduce.trim() || undefined,
-      expectedBehavior: newTicket.expectedBehavior.trim() || undefined,
-      actualBehavior: newTicket.actualBehavior.trim() || undefined,
-      severity: newTicket.severity,
-      environment: newTicket.environment,
-      testCaseId: newTicket.testCaseId.trim() || undefined,
-      regression: newTicket.regression,
-      workaround: newTicket.workaround.trim() || undefined
+      urlToPage: ticketData.urlToPage?.trim() || undefined,
+      browser: ticketData.browser?.trim() || undefined,
+      browserVersion: ticketData.browserVersion?.trim() || undefined,
+      operatingSystem: ticketData.operatingSystem?.trim() || undefined,
+      deviceType: ticketData.deviceType,
+      screenResolution: ticketData.screenResolution?.trim() || undefined,
+      stepsToReproduce: ticketData.stepsToReproduce?.trim() || undefined,
+      expectedBehavior: ticketData.expectedBehavior?.trim() || undefined,
+      actualBehavior: ticketData.actualBehavior?.trim() || undefined,
+      severity: ticketData.severity,
+      environment: ticketData.environment,
+      testCaseId: ticketData.testCaseId?.trim() || undefined,
+      regression: ticketData.regression,
+      workaround: ticketData.workaround?.trim() || undefined
     };
 
     // Add optimistic ticket to the list
     setTickets(prevTickets => [optimisticTicket, ...prevTickets]);
 
-    // Close modal and reset form immediately
+    // Close modal immediately
     setShowCreateModal(false);
-    setNewTicket({
-      title: '',
-      description: '',
-      category: 'bug',
-      priority: 'medium',
-      tags: [],
-      tagInput: '',
-      attachments: [],
-      // Testing-specific fields
-      urlToPage: '',
-      browser: '',
-      browserVersion: '',
-      operatingSystem: '',
-      deviceType: 'desktop',
-      screenResolution: '',
-      stepsToReproduce: '',
-      expectedBehavior: '',
-      actualBehavior: '',
-      severity: 'minor',
-      environment: 'production',
-      testCaseId: '',
-      regression: false,
-      workaround: ''
-    });
 
     try {
       const ticketId = await FeedbackService.createTicket(
-        {
-          title: newTicket.title.trim(),
-          description: newTicket.description.trim(),
-          category: newTicket.category,
-          priority: newTicket.priority,
-          tags: newTicket.tags,
-          attachments: newTicket.attachments,
-          // Testing-specific fields
-          urlToPage: newTicket.urlToPage.trim() || undefined,
-          browser: newTicket.browser.trim() || undefined,
-          browserVersion: newTicket.browserVersion.trim() || undefined,
-          operatingSystem: newTicket.operatingSystem.trim() || undefined,
-          deviceType: newTicket.deviceType,
-          screenResolution: newTicket.screenResolution.trim() || undefined,
-          stepsToReproduce: newTicket.stepsToReproduce.trim() || undefined,
-          expectedBehavior: newTicket.expectedBehavior.trim() || undefined,
-          actualBehavior: newTicket.actualBehavior.trim() || undefined,
-          severity: newTicket.severity,
-          environment: newTicket.environment,
-          testCaseId: newTicket.testCaseId.trim() || undefined,
-          regression: newTicket.regression,
-          workaround: newTicket.workaround.trim() || undefined
-        },
+        ticketData,
         userProfile?.uid || '',
         userProfile?.displayName || 'Unknown User',
         userProfile?.email || ''
@@ -425,50 +330,9 @@ export default function AdminTestingFeedback() {
     }
   };
 
-  const handleAddTag = () => {
-    const tag = newTicket.tagInput.trim();
-    if (tag && !newTicket.tags.includes(tag)) {
-      setNewTicket(prev => ({
-        ...prev,
-        tags: [...prev.tags, tag],
-        tagInput: ''
-      }));
-    }
-  };
-
-  const handleRemoveTag = (tagToRemove: string) => {
-    setNewTicket(prev => ({
-      ...prev,
-      tags: prev.tags.filter(tag => tag !== tagToRemove)
-    }));
-  };
-
-  const handleFileSelect = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(event.target.files || []);
-    setNewTicket(prev => ({
-      ...prev,
-      attachments: [...prev.attachments, ...files]
-    }));
-  };
-
-  const handleRemoveFile = (indexToRemove: number) => {
-    setNewTicket(prev => ({
-      ...prev,
-      attachments: prev.attachments.filter((_, index) => index !== indexToRemove)
-    }));
-  };
-
-  const formatFileSize = (bytes: number) => {
-    if (bytes === 0) return '0 Bytes';
-    const k = 1024;
-    const sizes = ['Bytes', 'KB', 'MB', 'GB'];
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
-  };
-
   const handleViewTicket = (ticket: FeedbackTicket) => {
-    setSelectedTicket(ticket);
-    setShowTicketModal(true);
+    setViewingTicket(ticket);
+    setShowViewModal(true);
   };
 
   const handleEditTicket = (ticket: FeedbackTicket) => {
@@ -476,17 +340,19 @@ export default function AdminTestingFeedback() {
     setShowEditModal(true);
   };
 
-  const handleUpdateTicket = async () => {
+  const handleUpdateTicket = async (ticketData: Partial<FeedbackTicket>) => {
     if (!editingTicket) return;
 
     const originalTicket = tickets.find(t => t.id === editingTicket.id);
     if (!originalTicket) return;
 
+    const updatedTicket = { ...editingTicket, ...ticketData };
+
     // Optimistic update
     setTickets(prevTickets => 
       prevTickets.map(ticket => 
         ticket.id === editingTicket.id 
-          ? { ...ticket, ...editingTicket, updatedAt: new Date() }
+          ? { ...ticket, ...updatedTicket, updatedAt: new Date() }
           : ticket
       )
     );
@@ -497,27 +363,27 @@ export default function AdminTestingFeedback() {
 
     try {
       await FeedbackService.updateTicket(editingTicket.id, {
-        title: editingTicket.title,
-        description: editingTicket.description,
-        category: editingTicket.category,
-        priority: editingTicket.priority,
-        status: editingTicket.status,
-        tags: editingTicket.tags,
+        title: updatedTicket.title,
+        description: updatedTicket.description,
+        category: updatedTicket.category,
+        priority: updatedTicket.priority,
+        status: updatedTicket.status,
+        tags: updatedTicket.tags,
         // Testing-specific fields
-        urlToPage: editingTicket.urlToPage,
-        browser: editingTicket.browser,
-        browserVersion: editingTicket.browserVersion,
-        operatingSystem: editingTicket.operatingSystem,
-        deviceType: editingTicket.deviceType,
-        screenResolution: editingTicket.screenResolution,
-        stepsToReproduce: editingTicket.stepsToReproduce,
-        expectedBehavior: editingTicket.expectedBehavior,
-        actualBehavior: editingTicket.actualBehavior,
-        severity: editingTicket.severity,
-        environment: editingTicket.environment,
-        testCaseId: editingTicket.testCaseId,
-        regression: editingTicket.regression,
-        workaround: editingTicket.workaround
+        urlToPage: updatedTicket.urlToPage,
+        browser: updatedTicket.browser,
+        browserVersion: updatedTicket.browserVersion,
+        operatingSystem: updatedTicket.operatingSystem,
+        deviceType: updatedTicket.deviceType,
+        screenResolution: updatedTicket.screenResolution,
+        stepsToReproduce: updatedTicket.stepsToReproduce,
+        expectedBehavior: updatedTicket.expectedBehavior,
+        actualBehavior: updatedTicket.actualBehavior,
+        severity: updatedTicket.severity,
+        environment: updatedTicket.environment,
+        testCaseId: updatedTicket.testCaseId,
+        regression: updatedTicket.regression,
+        workaround: updatedTicket.workaround
       });
     } catch (err) {
       console.error('Error updating ticket:', err);
@@ -570,28 +436,18 @@ export default function AdminTestingFeedback() {
     setShowCommentsSidebar(true);
   };
 
-  const handleCloseComments = () => {
-    setShowCommentsSidebar(false);
-    setSelectedTicketForComments(null);
-    setNewComment({
-      content: '',
-      isInternal: false,
-      attachments: []
-    });
-  };
-
-  const handleAddComment = async () => {
-    if (!selectedTicketForComments || !newComment.content.trim()) return;
+  const handleAddComment = async (content: string, isInternal: boolean, attachments: File[]) => {
+    if (!selectedTicketForComments) return;
 
     const optimisticComment = {
       id: `temp-${Date.now()}`,
       ticketId: selectedTicketForComments.id,
       authorId: userProfile?.uid || '',
       authorName: userProfile?.displayName || 'Unknown User',
-      content: newComment.content.trim(),
+      content: content.trim(),
       createdAt: new Date(),
       updatedAt: new Date(),
-      isInternal: newComment.isInternal,
+      isInternal: isInternal,
       attachments: []
     };
 
@@ -617,18 +473,11 @@ export default function AdminTestingFeedback() {
       )
     );
 
-    // Clear comment form immediately
-    setNewComment({
-      content: '',
-      isInternal: false,
-      attachments: []
-    });
-
     try {
       const commentId = await FeedbackService.addComment(selectedTicketForComments.id, {
-        content: newComment.content.trim(),
-        isInternal: newComment.isInternal,
-        attachments: newComment.attachments
+        content: content.trim(),
+        isInternal: isInternal,
+        attachments: attachments
       }, userProfile?.uid || '', userProfile?.displayName || 'Unknown User');
 
       // Replace optimistic comment with real one
@@ -678,21 +527,6 @@ export default function AdminTestingFeedback() {
         )
       );
     }
-  };
-
-  const handleCommentFileSelect = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const files = Array.from(event.target.files || []);
-    setNewComment(prev => ({
-      ...prev,
-      attachments: [...prev.attachments, ...files]
-    }));
-  };
-
-  const handleRemoveCommentFile = (indexToRemove: number) => {
-    setNewComment(prev => ({
-      ...prev,
-      attachments: prev.attachments.filter((_, index) => index !== indexToRemove)
-    }));
   };
 
   const formatDate = (date: Date) => {
@@ -1095,1150 +929,40 @@ export default function AdminTestingFeedback() {
         </div>
       )}
 
-      {/* Create Ticket Modal */}
-      {showCreateModal && (
-        <div className="modal-overlay" onClick={() => setShowCreateModal(false)}>
-          <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-            <div className="modal-header">
-              <h3>Create New Ticket</h3>
-              <button 
-                className="modal-close"
-                onClick={() => setShowCreateModal(false)}
-                title="Close modal"
-                aria-label="Close modal"
-              >
-                <FaTimes />
-              </button>
-            </div>
-            
-            <div className="modal-body">
-              <div className="form-group">
-                <label htmlFor="ticket-title">Title *</label>
-                <input
-                  id="ticket-title"
-                  type="text"
-                  value={newTicket.title}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, title: e.target.value }))}
-                  placeholder="Brief description of the issue"
-                  className="form-input"
-                />
-              </div>
-              
-              <div className="form-group">
-                <label htmlFor="ticket-description">Description *</label>
-                <textarea
-                  id="ticket-description"
-                  value={newTicket.description}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, description: e.target.value }))}
-                  placeholder="Detailed description of the issue or feature request"
-                  className="form-textarea"
-                  rows={4}
-                />
-              </div>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="ticket-category">Category</label>
-                  <select
-                    id="ticket-category"
-                    value={newTicket.category}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, category: e.target.value as FeedbackCategory }))}
-                    className="form-select"
-                  >
-                    {CATEGORY_OPTIONS.map(category => (
-                      <option key={category} value={category}>
-                        {category.replace('_', ' ').toUpperCase()}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="ticket-priority">Priority</label>
-                  <select
-                    id="ticket-priority"
-                    value={newTicket.priority}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, priority: e.target.value as FeedbackPriority }))}
-                    className="form-select"
-                  >
-                    {PRIORITY_OPTIONS.map(priority => (
-                      <option key={priority} value={priority}>
-                        {priority.toUpperCase()}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-              </div>
-              
-              <div className="form-group">
-                <label htmlFor="ticket-tags">Tags</label>
-                <div className="tag-input-container">
-                  <input
-                    id="ticket-tags"
-                    type="text"
-                    value={newTicket.tagInput}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, tagInput: e.target.value }))}
-                    onKeyPress={(e) => e.key === 'Enter' && handleAddTag()}
-                    placeholder="Add a tag and press Enter"
-                    className="form-input"
-                  />
-                  <button
-                    type="button"
-                    onClick={handleAddTag}
-                    className="tag-add-btn"
-                    title="Add tag"
-                    aria-label="Add tag"
-                  >
-                    <FaPlus />
-                  </button>
-                </div>
-                {newTicket.tags.length > 0 && (
-                  <div className="tag-list">
-                    {newTicket.tags.map(tag => (
-                      <span key={tag} className="tag">
-                        <FaTag /> {tag}
-                        <button
-                          type="button"
-                          onClick={() => handleRemoveTag(tag)}
-                          className="tag-remove"
-                          title="Remove tag"
-                          aria-label={`Remove ${tag} tag`}
-                        >
-                          <FaTimes />
-                        </button>
-                      </span>
-                    ))}
-                  </div>
-                )}
-              </div>
+      {/* Modals */}
+      <CreateTicketModal
+        isOpen={showCreateModal}
+        onClose={() => setShowCreateModal(false)}
+        onSubmit={handleCreateTicket}
+      />
 
-              <div className="form-group">
-                <label htmlFor="ticket-attachments">Attachments</label>
-                <div className="file-input-container">
-                  <input
-                    id="ticket-attachments"
-                    type="file"
-                    multiple
-                    onChange={handleFileSelect}
-                    className="file-input"
-                    accept="image/*,video/*,.pdf,.doc,.docx,.txt,.csv,.xlsx,.zip"
-                  />
-                  <label htmlFor="ticket-attachments" className="file-input-label">
-                    <FaPlus /> Choose Files
-                  </label>
-                  <span className="file-input-hint">
-                    Images, videos, documents (max 10MB each)
-                  </span>
-                </div>
-                {newTicket.attachments.length > 0 && (
-                  <div className="file-list">
-                    {newTicket.attachments.map((file, index) => (
-                      <div key={index} className="file-item">
-                        <div className="file-info">
-                          <span className="file-name">{file.name}</span>
-                          <span className="file-size">{formatFileSize(file.size)}</span>
-                        </div>
-                        <button
-                          type="button"
-                          onClick={() => handleRemoveFile(index)}
-                          className="file-remove"
-                          title="Remove file"
-                          aria-label={`Remove ${file.name}`}
-                        >
-                          <FaTimes />
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                )}
-              </div>
+      <ViewTicketModal
+        isOpen={showViewModal}
+        ticket={viewingTicket}
+        onClose={() => setShowViewModal(false)}
+        onEdit={handleEditTicket}
+      />
 
-              {/* Testing-Specific Fields */}
-              <div className="form-section-divider">
-                <h4>Testing Information</h4>
-              </div>
+      <EditTicketModal
+        isOpen={showEditModal}
+        ticket={editingTicket}
+        onClose={() => setShowEditModal(false)}
+        onUpdate={handleUpdateTicket}
+      />
 
-              <div className="form-group">
-                <label htmlFor="ticket-url">URL to Page</label>
-                <input
-                  id="ticket-url"
-                  type="url"
-                  value={newTicket.urlToPage}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, urlToPage: e.target.value }))}
-                  placeholder="https://example.com/page"
-                  className="form-input"
-                />
-              </div>
+      <DeleteTicketModal
+        isOpen={showDeleteModal}
+        ticket={deletingTicket}
+        onClose={() => setShowDeleteModal(false)}
+        onConfirm={handleConfirmDelete}
+      />
 
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="ticket-browser">Browser</label>
-                  <select
-                    id="ticket-browser"
-                    value={newTicket.browser}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, browser: e.target.value }))}
-                    className="form-select"
-                  >
-                    <option value="">Select Browser</option>
-                    <option value="Chrome">Chrome</option>
-                    <option value="Firefox">Firefox</option>
-                    <option value="Safari">Safari</option>
-                    <option value="Edge">Edge</option>
-                    <option value="Opera">Opera</option>
-                    <option value="Other">Other</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="ticket-browser-version">Browser Version</label>
-                  <input
-                    id="ticket-browser-version"
-                    type="text"
-                    value={newTicket.browserVersion}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, browserVersion: e.target.value }))}
-                    placeholder="e.g., 120.0.6099.109"
-                    className="form-input"
-                  />
-                </div>
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="ticket-os">Operating System</label>
-                  <select
-                    id="ticket-os"
-                    value={newTicket.operatingSystem}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, operatingSystem: e.target.value }))}
-                    className="form-select"
-                  >
-                    <option value="">Select OS</option>
-                    <option value="Windows 11">Windows 11</option>
-                    <option value="Windows 10">Windows 10</option>
-                    <option value="macOS">macOS</option>
-                    <option value="Linux">Linux</option>
-                    <option value="iOS">iOS</option>
-                    <option value="Android">Android</option>
-                    <option value="Other">Other</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="ticket-device">Device Type</label>
-                  <select
-                    id="ticket-device"
-                    value={newTicket.deviceType}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, deviceType: e.target.value as 'desktop' | 'mobile' | 'tablet' }))}
-                    className="form-select"
-                  >
-                    <option value="desktop">Desktop</option>
-                    <option value="mobile">Mobile</option>
-                    <option value="tablet">Tablet</option>
-                  </select>
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-resolution">Screen Resolution</label>
-                <input
-                  id="ticket-resolution"
-                  type="text"
-                  value={newTicket.screenResolution}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, screenResolution: e.target.value }))}
-                  placeholder="e.g., 1920x1080"
-                  className="form-input"
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-steps">Steps to Reproduce</label>
-                <textarea
-                  id="ticket-steps"
-                  value={newTicket.stepsToReproduce}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, stepsToReproduce: e.target.value }))}
-                  placeholder="1. Go to the page...&#10;2. Click on the button...&#10;3. Observe the issue..."
-                  className="form-textarea"
-                  rows={3}
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-expected">Expected Behavior</label>
-                <textarea
-                  id="ticket-expected"
-                  value={newTicket.expectedBehavior}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, expectedBehavior: e.target.value }))}
-                  placeholder="What should happen when following the steps?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-actual">Actual Behavior</label>
-                <textarea
-                  id="ticket-actual"
-                  value={newTicket.actualBehavior}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, actualBehavior: e.target.value }))}
-                  placeholder="What actually happens instead?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="ticket-severity">Severity</label>
-                  <select
-                    id="ticket-severity"
-                    value={newTicket.severity}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, severity: e.target.value as 'cosmetic' | 'minor' | 'major' | 'critical' | 'blocker' }))}
-                    className="form-select"
-                  >
-                    <option value="cosmetic">Cosmetic</option>
-                    <option value="minor">Minor</option>
-                    <option value="major">Major</option>
-                    <option value="critical">Critical</option>
-                    <option value="blocker">Blocker</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="ticket-environment">Environment</label>
-                  <select
-                    id="ticket-environment"
-                    value={newTicket.environment}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, environment: e.target.value as 'development' | 'staging' | 'production' }))}
-                    className="form-select"
-                  >
-                    <option value="development">Development</option>
-                    <option value="staging">Staging</option>
-                    <option value="production">Production</option>
-                  </select>
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-testcase">Test Case ID</label>
-                <input
-                  id="ticket-testcase"
-                  type="text"
-                  value={newTicket.testCaseId}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, testCaseId: e.target.value }))}
-                  placeholder="e.g., TC-001, Test-123"
-                  className="form-input"
-                />
-              </div>
-
-              <div className="form-group">
-                <label className="checkbox-label">
-                  <input
-                    type="checkbox"
-                    checked={newTicket.regression}
-                    onChange={(e) => setNewTicket(prev => ({ ...prev, regression: e.target.checked }))}
-                    className="checkbox-input"
-                  />
-                  <span className="checkbox-text">This is a regression bug</span>
-                </label>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="ticket-workaround">Workaround</label>
-                <textarea
-                  id="ticket-workaround"
-                  value={newTicket.workaround}
-                  onChange={(e) => setNewTicket(prev => ({ ...prev, workaround: e.target.value }))}
-                  placeholder="Any temporary workaround or solution?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-            </div>
-            
-            <div className="modal-footer">
-              <button
-                className="btn btn-secondary"
-                onClick={() => setShowCreateModal(false)}
-              >
-                Cancel
-              </button>
-              <button
-                className="btn btn-primary"
-                onClick={handleCreateTicket}
-              >
-                <FaPlus /> Create Ticket
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* View Ticket Modal */}
-      {showTicketModal && selectedTicket && (
-        <div className="modal-overlay" onClick={() => setShowTicketModal(false)}>
-          <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-            <div className="modal-header">
-              <h3>Ticket Details</h3>
-              <button 
-                className="modal-close"
-                onClick={() => setShowTicketModal(false)}
-                title="Close modal"
-                aria-label="Close modal"
-              >
-                <FaTimes />
-              </button>
-            </div>
-            
-            <div className="modal-body">
-              <div className="ticket-detail-section">
-                <h4>Title</h4>
-                <p className="ticket-detail-value">{selectedTicket.title}</p>
-              </div>
-              
-              <div className="ticket-detail-section">
-                <h4>Description</h4>
-                <p className="ticket-detail-value">{selectedTicket.description}</p>
-              </div>
-              
-              <div className="ticket-detail-row">
-                <div className="ticket-detail-section">
-                  <h4>Category</h4>
-                  <span className="ticket-detail-badge category">
-                    {selectedTicket.category.replace('_', ' ').toUpperCase()}
-                  </span>
-                </div>
-                
-                <div className="ticket-detail-section">
-                  <h4>Priority</h4>
-                  <span 
-                    className="ticket-detail-badge priority"
-                    style={{ backgroundColor: PRIORITY_COLORS[selectedTicket.priority] }}
-                  >
-                    {selectedTicket.priority.toUpperCase()}
-                  </span>
-                </div>
-              </div>
-              
-              <div className="ticket-detail-section">
-                <h4>Status</h4>
-                <span 
-                  className="ticket-detail-badge status"
-                  style={{ backgroundColor: STATUS_COLORS[selectedTicket.status] }}
-                >
-                  {getStatusIcon(selectedTicket.status)}
-                  {selectedTicket.status.replace('_', ' ').toUpperCase()}
-                </span>
-              </div>
-              
-              {selectedTicket.tags.length > 0 && (
-                <div className="ticket-detail-section">
-                  <h4>Tags</h4>
-                  <div className="ticket-detail-tags">
-                    {selectedTicket.tags.map(tag => (
-                      <span key={tag} className="ticket-detail-tag">
-                        <FaTag /> {tag}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              )}
-              
-              {selectedTicket.attachments && selectedTicket.attachments.length > 0 && (
-                <div className="ticket-detail-section">
-                  <h4>Attachments</h4>
-                  <div className="ticket-detail-attachments">
-                    {selectedTicket.attachments.map(attachment => (
-                      <div key={attachment.id} className="ticket-detail-attachment">
-                        <span className="attachment-name">{attachment.name}</span>
-                        <span className="attachment-size">{formatFileSize(attachment.size)}</span>
-                        <a 
-                          href={attachment.url} 
-                          target="_blank" 
-                          rel="noopener noreferrer"
-                          className="attachment-link"
-                        >
-                          View
-                        </a>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-              )}
-              
-              <div className="ticket-detail-row">
-                <div className="ticket-detail-section">
-                  <h4>Created</h4>
-                  <p className="ticket-detail-value">
-                    {formatDate(selectedTicket.createdAt)} by {selectedTicket.reporterName}
-                  </p>
-                </div>
-                
-                <div className="ticket-detail-section">
-                  <h4>Last Updated</h4>
-                  <p className="ticket-detail-value">
-                    {formatDate(selectedTicket.updatedAt)}
-                  </p>
-                </div>
-              </div>
-              
-              {/* Testing-Specific Fields */}
-              {selectedTicket.urlToPage && (
-                <div className="ticket-detail-section">
-                  <h4>URL to Page</h4>
-                  <p className="ticket-detail-value">
-                    <a href={selectedTicket.urlToPage} target="_blank" rel="noopener noreferrer" className="attachment-link">
-                      {selectedTicket.urlToPage}
-                    </a>
-                  </p>
-                </div>
-              )}
-
-              {(selectedTicket.browser || selectedTicket.browserVersion || selectedTicket.operatingSystem) && (
-                <div className="ticket-detail-row">
-                  {selectedTicket.browser && (
-                    <div className="ticket-detail-section">
-                      <h4>Browser</h4>
-                      <p className="ticket-detail-value">{selectedTicket.browser}</p>
-                    </div>
-                  )}
-                  
-                  {selectedTicket.browserVersion && (
-                    <div className="ticket-detail-section">
-                      <h4>Browser Version</h4>
-                      <p className="ticket-detail-value">{selectedTicket.browserVersion}</p>
-                    </div>
-                  )}
-                </div>
-              )}
-
-              {(selectedTicket.operatingSystem || selectedTicket.deviceType || selectedTicket.screenResolution) && (
-                <div className="ticket-detail-row">
-                  {selectedTicket.operatingSystem && (
-                    <div className="ticket-detail-section">
-                      <h4>Operating System</h4>
-                      <p className="ticket-detail-value">{selectedTicket.operatingSystem}</p>
-                    </div>
-                  )}
-                  
-                  <div className="ticket-detail-section">
-                    <h4>Device Type</h4>
-                    <p className="ticket-detail-value">{selectedTicket.deviceType}</p>
-                  </div>
-                </div>
-              )}
-
-              {selectedTicket.screenResolution && (
-                <div className="ticket-detail-section">
-                  <h4>Screen Resolution</h4>
-                  <p className="ticket-detail-value">{selectedTicket.screenResolution}</p>
-                </div>
-              )}
-
-              {selectedTicket.stepsToReproduce && (
-                <div className="ticket-detail-section">
-                  <h4>Steps to Reproduce</h4>
-                  <p className="ticket-detail-value" style={{ whiteSpace: 'pre-wrap' }}>{selectedTicket.stepsToReproduce}</p>
-                </div>
-              )}
-
-              {selectedTicket.expectedBehavior && (
-                <div className="ticket-detail-section">
-                  <h4>Expected Behavior</h4>
-                  <p className="ticket-detail-value" style={{ whiteSpace: 'pre-wrap' }}>{selectedTicket.expectedBehavior}</p>
-                </div>
-              )}
-
-              {selectedTicket.actualBehavior && (
-                <div className="ticket-detail-section">
-                  <h4>Actual Behavior</h4>
-                  <p className="ticket-detail-value" style={{ whiteSpace: 'pre-wrap' }}>{selectedTicket.actualBehavior}</p>
-                </div>
-              )}
-
-              {(selectedTicket.severity || selectedTicket.environment) && (
-                <div className="ticket-detail-row">
-                  <div className="ticket-detail-section">
-                    <h4>Severity</h4>
-                    <span className="ticket-detail-badge" style={{ backgroundColor: PRIORITY_COLORS[selectedTicket.severity as FeedbackPriority] }}>
-                      {selectedTicket.severity?.toUpperCase()}
-                    </span>
-                  </div>
-                  
-                  <div className="ticket-detail-section">
-                    <h4>Environment</h4>
-                    <span className="ticket-detail-badge" style={{ backgroundColor: '#6b7280' }}>
-                      {selectedTicket.environment?.toUpperCase()}
-                    </span>
-                  </div>
-                </div>
-              )}
-
-              {selectedTicket.testCaseId && (
-                <div className="ticket-detail-section">
-                  <h4>Test Case ID</h4>
-                  <p className="ticket-detail-value">{selectedTicket.testCaseId}</p>
-                </div>
-              )}
-
-              <div className="ticket-detail-section">
-                <h4>Regression Bug</h4>
-                <p className="ticket-detail-value">
-                  {selectedTicket.regression ? (
-                    <span style={{ color: '#dc2626', fontWeight: '600' }}>Yes</span>
-                  ) : (
-                    <span style={{ color: '#059669', fontWeight: '600' }}>No</span>
-                  )}
-                </p>
-              </div>
-
-              {selectedTicket.workaround && (
-                <div className="ticket-detail-section">
-                  <h4>Workaround</h4>
-                  <p className="ticket-detail-value" style={{ whiteSpace: 'pre-wrap' }}>{selectedTicket.workaround}</p>
-                </div>
-              )}
-              
-              <div className="ticket-detail-section">
-                <h4>Votes</h4>
-                <div className="ticket-detail-votes">
-                  <span className="vote-count">{selectedTicket.votes}</span>
-                  <span className="vote-label">votes</span>
-                </div>
-              </div>
-            </div>
-            
-            <div className="modal-footer">
-              <button
-                className="btn btn-secondary"
-                onClick={() => setShowTicketModal(false)}
-              >
-                Close
-              </button>
-              <button
-                className="btn btn-primary"
-                onClick={() => {
-                  setShowTicketModal(false);
-                  handleEditTicket(selectedTicket);
-                }}
-              >
-                <FaEdit /> Edit Ticket
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Edit Ticket Modal */}
-      {showEditModal && editingTicket && (
-        <div className="modal-overlay" onClick={() => setShowEditModal(false)}>
-          <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-            <div className="modal-header">
-              <h3>Edit Ticket</h3>
-              <button 
-                className="modal-close"
-                onClick={() => setShowEditModal(false)}
-                title="Close modal"
-                aria-label="Close modal"
-              >
-                <FaTimes />
-              </button>
-            </div>
-            
-            <div className="modal-body">
-              <div className="form-group">
-                <label htmlFor="edit-ticket-title">Title *</label>
-                <input
-                  id="edit-ticket-title"
-                  type="text"
-                  value={editingTicket.title}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, title: e.target.value } : null)}
-                  placeholder="Brief description of the issue"
-                  className="form-input"
-                />
-              </div>
-              
-              <div className="form-group">
-                <label htmlFor="edit-ticket-description">Description *</label>
-                <textarea
-                  id="edit-ticket-description"
-                  value={editingTicket.description}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, description: e.target.value } : null)}
-                  placeholder="Detailed description of the issue or feature request"
-                  className="form-textarea"
-                  rows={4}
-                />
-              </div>
-              
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-category">Category</label>
-                  <select
-                    id="edit-ticket-category"
-                    value={editingTicket.category}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, category: e.target.value as FeedbackCategory } : null)}
-                    className="form-select"
-                  >
-                    {CATEGORY_OPTIONS.map(category => (
-                      <option key={category} value={category}>
-                        {category.replace('_', ' ').toUpperCase()}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-priority">Priority</label>
-                  <select
-                    id="edit-ticket-priority"
-                    value={editingTicket.priority}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, priority: e.target.value as FeedbackPriority } : null)}
-                    className="form-select"
-                  >
-                    {PRIORITY_OPTIONS.map(priority => (
-                      <option key={priority} value={priority}>
-                        {priority.toUpperCase()}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-              </div>
-              
-              <div className="form-group">
-                <label htmlFor="edit-ticket-status">Status</label>
-                <select
-                  id="edit-ticket-status"
-                  value={editingTicket.status}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, status: e.target.value as FeedbackStatus } : null)}
-                  className="form-select"
-                >
-                  {STATUS_OPTIONS.map(status => (
-                    <option key={status} value={status}>
-                      {status.replace('_', ' ').toUpperCase()}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Testing-Specific Fields */}
-              <div className="form-section-divider">
-                <h4>Testing Information</h4>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-url">URL to Page</label>
-                <input
-                  id="edit-ticket-url"
-                  type="url"
-                  value={editingTicket.urlToPage || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, urlToPage: e.target.value } : null)}
-                  placeholder="https://example.com/page"
-                  className="form-input"
-                />
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-browser">Browser</label>
-                  <select
-                    id="edit-ticket-browser"
-                    value={editingTicket.browser || ''}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, browser: e.target.value } : null)}
-                    className="form-select"
-                  >
-                    <option value="">Select Browser</option>
-                    <option value="Chrome">Chrome</option>
-                    <option value="Firefox">Firefox</option>
-                    <option value="Safari">Safari</option>
-                    <option value="Edge">Edge</option>
-                    <option value="Opera">Opera</option>
-                    <option value="Other">Other</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-browser-version">Browser Version</label>
-                  <input
-                    id="edit-ticket-browser-version"
-                    type="text"
-                    value={editingTicket.browserVersion || ''}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, browserVersion: e.target.value } : null)}
-                    placeholder="e.g., 120.0.6099.109"
-                    className="form-input"
-                  />
-                </div>
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-os">Operating System</label>
-                  <select
-                    id="edit-ticket-os"
-                    value={editingTicket.operatingSystem || ''}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, operatingSystem: e.target.value } : null)}
-                    className="form-select"
-                  >
-                    <option value="">Select OS</option>
-                    <option value="Windows 11">Windows 11</option>
-                    <option value="Windows 10">Windows 10</option>
-                    <option value="macOS">macOS</option>
-                    <option value="Linux">Linux</option>
-                    <option value="iOS">iOS</option>
-                    <option value="Android">Android</option>
-                    <option value="Other">Other</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-device">Device Type</label>
-                  <select
-                    id="edit-ticket-device"
-                    value={editingTicket.deviceType}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, deviceType: e.target.value as 'desktop' | 'mobile' | 'tablet' } : null)}
-                    className="form-select"
-                  >
-                    <option value="desktop">Desktop</option>
-                    <option value="mobile">Mobile</option>
-                    <option value="tablet">Tablet</option>
-                  </select>
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-resolution">Screen Resolution</label>
-                <input
-                  id="edit-ticket-resolution"
-                  type="text"
-                  value={editingTicket.screenResolution || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, screenResolution: e.target.value } : null)}
-                  placeholder="e.g., 1920x1080"
-                  className="form-input"
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-steps">Steps to Reproduce</label>
-                <textarea
-                  id="edit-ticket-steps"
-                  value={editingTicket.stepsToReproduce || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, stepsToReproduce: e.target.value } : null)}
-                  placeholder="1. Go to the page...&#10;2. Click on the button...&#10;3. Observe the issue..."
-                  className="form-textarea"
-                  rows={3}
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-expected">Expected Behavior</label>
-                <textarea
-                  id="edit-ticket-expected"
-                  value={editingTicket.expectedBehavior || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, expectedBehavior: e.target.value } : null)}
-                  placeholder="What should happen when following the steps?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-actual">Actual Behavior</label>
-                <textarea
-                  id="edit-ticket-actual"
-                  value={editingTicket.actualBehavior || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, actualBehavior: e.target.value } : null)}
-                  placeholder="What actually happens instead?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-
-              <div className="form-row">
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-severity">Severity</label>
-                  <select
-                    id="edit-ticket-severity"
-                    value={editingTicket.severity}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, severity: e.target.value as 'cosmetic' | 'minor' | 'major' | 'critical' | 'blocker' } : null)}
-                    className="form-select"
-                  >
-                    <option value="cosmetic">Cosmetic</option>
-                    <option value="minor">Minor</option>
-                    <option value="major">Major</option>
-                    <option value="critical">Critical</option>
-                    <option value="blocker">Blocker</option>
-                  </select>
-                </div>
-                
-                <div className="form-group">
-                  <label htmlFor="edit-ticket-environment">Environment</label>
-                  <select
-                    id="edit-ticket-environment"
-                    value={editingTicket.environment}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, environment: e.target.value as 'development' | 'staging' | 'production' } : null)}
-                    className="form-select"
-                  >
-                    <option value="development">Development</option>
-                    <option value="staging">Staging</option>
-                    <option value="production">Production</option>
-                  </select>
-                </div>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-testcase">Test Case ID</label>
-                <input
-                  id="edit-ticket-testcase"
-                  type="text"
-                  value={editingTicket.testCaseId || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, testCaseId: e.target.value } : null)}
-                  placeholder="e.g., TC-001, Test-123"
-                  className="form-input"
-                />
-              </div>
-
-              <div className="form-group">
-                <label className="checkbox-label">
-                  <input
-                    type="checkbox"
-                    checked={editingTicket.regression}
-                    onChange={(e) => setEditingTicket(prev => prev ? { ...prev, regression: e.target.checked } : null)}
-                    className="checkbox-input"
-                  />
-                  <span className="checkbox-text">This is a regression bug</span>
-                </label>
-              </div>
-
-              <div className="form-group">
-                <label htmlFor="edit-ticket-workaround">Workaround</label>
-                <textarea
-                  id="edit-ticket-workaround"
-                  value={editingTicket.workaround || ''}
-                  onChange={(e) => setEditingTicket(prev => prev ? { ...prev, workaround: e.target.value } : null)}
-                  placeholder="Any temporary workaround or solution?"
-                  className="form-textarea"
-                  rows={2}
-                />
-              </div>
-            </div>
-            
-            <div className="modal-footer">
-              <button
-                className="btn btn-secondary"
-                onClick={() => setShowEditModal(false)}
-              >
-                Cancel
-              </button>
-              <button
-                className="btn btn-primary"
-                onClick={handleUpdateTicket}
-              >
-                <FaEdit /> Update Ticket
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Delete Confirmation Modal */}
-      {showDeleteModal && deletingTicket && (
-        <div className="modal-overlay" onClick={handleCancelDelete}>
-          <div className="modal-content delete-modal" onClick={(e) => e.stopPropagation()}>
-            <div className="modal-header">
-              <h3>Delete Ticket</h3>
-              <button 
-                className="modal-close"
-                onClick={handleCancelDelete}
-                title="Close modal"
-                aria-label="Close modal"
-              >
-                <FaTimes />
-              </button>
-            </div>
-            
-            <div className="modal-body">
-              <div className="delete-warning">
-                <FaExclamationTriangle className="warning-icon" />
-                <h4>Are you sure you want to delete this ticket?</h4>
-                <p>This action cannot be undone. The following ticket will be permanently deleted:</p>
-              </div>
-              
-              <div className="ticket-to-delete">
-                <div className="ticket-preview">
-                  <div className="ticket-preview-header">
-                    <span className="ticket-category">
-                      {deletingTicket.category.replace('_', ' ').toUpperCase()}
-                    </span>
-                    <span 
-                      className="ticket-priority"
-                      style={{ backgroundColor: PRIORITY_COLORS[deletingTicket.priority] }}
-                    >
-                      {deletingTicket.priority.toUpperCase()}
-                    </span>
-                  </div>
-                  <h4 className="ticket-title">{deletingTicket.title}</h4>
-                  <p className="ticket-description">
-                    {deletingTicket.description.length > 100 
-                      ? `${deletingTicket.description.substring(0, 100)}...` 
-                      : deletingTicket.description
-                    }
-                  </p>
-                  <div className="ticket-meta">
-                    <span className="ticket-reporter">by {deletingTicket.reporterName}</span>
-                    <span className="ticket-date">{formatDate(deletingTicket.createdAt)}</span>
-                  </div>
-                </div>
-              </div>
-              
-              <div className="delete-consequences">
-                <h5>This will also delete:</h5>
-                <ul>
-                  <li>All comments on this ticket</li>
-                  <li>All file attachments</li>
-                  <li>Vote history</li>
-                  <li>All related data</li>
-                </ul>
-              </div>
-            </div>
-            
-            <div className="modal-footer">
-              <button
-                className="btn btn-secondary"
-                onClick={handleCancelDelete}
-              >
-                Cancel
-              </button>
-              <button
-                className="btn btn-danger"
-                onClick={handleConfirmDelete}
-              >
-                <FaTrash /> Delete Permanently
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {/* Comments Sidebar */}
-      {showCommentsSidebar && selectedTicketForComments && (
-        <div className="comments-sidebar-overlay" onClick={handleCloseComments}>
-          <div className="comments-sidebar" onClick={(e) => e.stopPropagation()}>
-            <div className="comments-header">
-              <h3>Comments - {selectedTicketForComments.title}</h3>
-              <button 
-                className="comments-close"
-                onClick={handleCloseComments}
-                title="Close comments"
-                aria-label="Close comments"
-              >
-                <FaTimes />
-              </button>
-            </div>
-            
-            <div className="comments-content">
-              <div className="comments-list">
-                {selectedTicketForComments.comments.map(comment => (
-                  <div key={comment.id} className={`comment ${comment.isInternal ? 'internal' : 'public'}`}>
-                    <div className="comment-header">
-                      <div className="comment-author">
-                        <FaUser className="comment-author-icon" />
-                        <span className="comment-author-name">{comment.authorName}</span>
-                        {comment.isInternal && (
-                          <span className="comment-internal-badge">Internal</span>
-                        )}
-                      </div>
-                      <span className="comment-date">{formatDate(comment.createdAt)}</span>
-                    </div>
-                    <div className="comment-body">
-                      <p>{comment.content}</p>
-                      {comment.attachments && comment.attachments.length > 0 && (
-                        <div className="comment-attachments">
-                          {comment.attachments.map(attachment => (
-                            <a
-                              key={attachment.id}
-                              href={attachment.url}
-                              target="_blank"
-                              rel="noopener noreferrer"
-                              className="comment-attachment"
-                            >
-                              <FaTag /> {attachment.name}
-                            </a>
-                          ))}
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                ))}
-                {selectedTicketForComments.comments.length === 0 && (
-                  <div className="comments-empty">
-                    <FaComments className="comments-empty-icon" />
-                    <p>No comments yet. Be the first to comment!</p>
-                  </div>
-                )}
-              </div>
-              
-              <div className="comments-form">
-                <div className="comment-form-header">
-                  <h4>Add Comment</h4>
-                  <label className="comment-internal-toggle">
-                    <input
-                      type="checkbox"
-                      checked={newComment.isInternal}
-                      onChange={(e) => setNewComment(prev => ({ ...prev, isInternal: e.target.checked }))}
-                      className="checkbox-input"
-                    />
-                    <span className="checkbox-text">Internal comment</span>
-                  </label>
-                </div>
-                
-                <textarea
-                  value={newComment.content}
-                  onChange={(e) => setNewComment(prev => ({ ...prev, content: e.target.value }))}
-                  placeholder="Write your comment here..."
-                  className="comment-textarea"
-                  rows={4}
-                />
-                
-                <div className="comment-file-input">
-                  <input
-                    type="file"
-                    multiple
-                    onChange={handleCommentFileSelect}
-                    className="file-input"
-                    accept="image/*,video/*,.pdf,.doc,.docx,.txt,.csv,.xlsx,.zip"
-                  />
-                  <label className="file-input-label">
-                    <FaPlus /> Attach Files
-                  </label>
-                </div>
-                
-                {newComment.attachments.length > 0 && (
-                  <div className="comment-file-list">
-                    {newComment.attachments.map((file, index) => (
-                      <div key={index} className="comment-file-item">
-                        <span className="comment-file-name">{file.name}</span>
-                        <button
-                          type="button"
-                          onClick={() => handleRemoveCommentFile(index)}
-                          className="comment-file-remove"
-                          title="Remove file"
-                        >
-                          <FaTimes />
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                )}
-                
-                <button
-                  className="btn btn-primary comment-submit"
-                  onClick={handleAddComment}
-                  disabled={!newComment.content.trim()}
-                >
-                  <FaComments /> Add Comment
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      )}
+      <CommentsSidebar
+        isOpen={showCommentsSidebar}
+        ticket={selectedTicketForComments}
+        onClose={() => setShowCommentsSidebar(false)}
+        onAddComment={handleAddComment}
+      />
     </div>
   );
 }
